explicitly convert `MathLib::big{u}int` to/from other types / small cleanup (#7129)
fixes some compiler warnings with `USE_BOOST_INT128` as those implicit
conversions are not allowed

+++ lib/library.cpp
@@ -1055,7 +1055,7 @@ bool Library::isIntArgValid(const Token *ftok, int argnr, const MathLib::bigint 
     if (!ac || ac->valid.empty())
         return true;
     if (ac->valid.find('.') != std::string::npos)
-        return isFloatArgValid(ftok, argnr, argvalue);
+        return isFloatArgValid(ftok, argnr, static_cast<double>(argvalue));
     TokenList tokenList(nullptr);
     gettokenlistfromvalid(ac->valid, ftok->isCpp(), tokenList);
     for (const Token *tok = tokenList.front(); tok; tok = tok->next()) {
